lpms/requirements_dev.txt
Django==5.0.3
django-allauth==0.61.1
django-debug-toolbar==4.3.0
django-import-export==3.3.7
django-stubs==4.2.7
django-stubs-ext==4.2.7
flake8==7.0.0
flake8-print==5.0.0
Markdown==3.6
mypy==1.9.0
mypy-extensions==1.0.0
types-Markdown==3.6.0.20240316
python-dotenv==1.0.1
PyGithub==2.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
gunicorn==21.2.0
